Convert shift counts to int instead of long. Note that "int" means machine word (64 bits on a 64-bit machine), and "long" means Python bignum. Also note that int() actually converts to long if the value does not fit in an int; this means the behavior is unchanged.

+++ src/sage/rings/integer.pyx
@@ -1661,7 +1661,7 @@ cdef class Integer(sage.structure.element.EuclideanDomainElement):
             128
         """
         if PY_TYPE_CHECK(x, Integer) and isinstance(y, (Integer, int, long)):
-            return (<Integer>x)._lshift(long(y))
+            return (<Integer>x)._lshift(int(y))
         return bin_op(x, y, operator.lshift)
 
     cdef _rshift(Integer self, unsigned long int n):
@@ -1683,7 +1683,7 @@ cdef class Integer(sage.structure.element.EuclideanDomainElement):
             8
         """
         if PY_TYPE_CHECK(x, Integer) and isinstance(y, (Integer, int, long)):
-            return (<Integer>x)._rshift(long(y))
+            return (<Integer>x)._rshift(int(y))
         return bin_op(x, y, operator.rshift)
 
     cdef _and(Integer self, Integer other):
